requirements.txt
et-xmlfile==1.1.0
greenlet==3.0.1
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
playwright==1.40.0
pyee==11.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
